requirements.txt
aiodns==3.1.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
coverage==7.3.2
cryptography==41.0.5
cycler==0.12.1
flake8==6.1.0
fonttools==4.44.1
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
mccabe==0.7.0
multidict==6.0.4
numpy==1.26.1
packaging==23.2
pandas==2.1.3
patsy==0.5.3
Pillow==10.1.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pycares==4.4.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
pyparsing==3.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
SQLAlchemy==2.0.23
statsmodels==0.14.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.2
